Extract ScalaRewriteBase as an independent module.
